"Update CommonCheckTable component usage in Leads view and add searchDisplay state"
This commit updates the usage of the CommonCheckTable component in the Leads view to include the new searchDisplay and setSearchDisplay props. It also adds a new state variable, searchDisplay, to control the display of searched data. Additionally, the tableData prop is updated to use searchedDataOut when searchDisplay is true. The commit also includes a console log for searchedDataOut and updates the dependency array for the useEffect hook to include location.state.